% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MigratoryBirdsCount.R
\name{MigratoryBirdsCount_function}
\alias{MigratoryBirdsCount_function}
\title{A Global Migratory Birds Mapping Function}
\usage{
MigratoryBirdsCount_function(ans = TRUE)
}
\arguments{
\item{Yes}{Do you want to map migratory? Defaults to TRUE.}
}
\description{
This function allows you to visualize global migratory birds distribution
}
\examples{
MigratoryBirdsCount_function()
}
\keyword{MigratoryBirdsGlobal}
